requirements.txt
asgiref==3.4.1
astroid==2.4.2
certifi==2021.5.30
colorama==0.4.4
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.3.2
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
Pygments==2.10.0
pylint==2.6.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions @ file:///Users/ktietz/demo/mc3/conda-bld/typing_extensions_1629887998905/work
wincertstore==0.2
wrapt==1.12.1
